3tier-web-app-main/backend/requirements.txt
Flask==2.0.3
werkzeug==2.0.3
mysql-connector-python==8.0.31
flask-cors

